-add synced simulations -add some javadoc -unify some worker methods
