+++ src/main/java/org/javafling/pokerenlighter/simulation/worker/OmahaHiLoWorker.java
@@ -4,7 +4,6 @@ import org.javafling.pokerenlighter.combination.Card;
 import org.javafling.pokerenlighter.combination.Deck;
 import org.javafling.pokerenlighter.combination.OmahaCombination;
 import org.javafling.pokerenlighter.simulation.HandType;
-import org.javafling.pokerenlighter.simulation.PlayerProfile;
 import org.javafling.pokerenlighter.simulation.PokerType;
 
 public class OmahaHiLoWorker extends SimulationWorker
@@ -14,22 +13,7 @@ public class OmahaHiLoWorker extends SimulationWorker
         @Override
         public OmahaHiLoWorker build()
         {
-            if (super.getRounds() <= 0) {
-                throw new IllegalStateException("The number of rounds must be a strictly positive number");
-            } else if (super.getProfiles() == null || super.getProfiles().size() < 2) {
-                throw new IllegalStateException("There need to be at least 2 players in every simulation.");
-            } else if (super.getUpdateInterval() <= 0 || 100 % super.getUpdateInterval() != 0) {
-                throw new IllegalStateException("Invalid update interval value");
-            } else if (super.getNotifiable() == null) {
-                throw new IllegalStateException("There needs to be a notifiable for this worker");
-            }
-            
-            for (PlayerProfile profile : super.getProfiles()) {
-                if (profile == null) {
-                    throw new NullPointerException();
-                }
-            }
-            
+            validate();
             return new OmahaHiLoWorker(this);
         }
     }
@@ -165,20 +149,9 @@ public class OmahaHiLoWorker extends SimulationWorker
                     this.loses[i]++;
                 }
             }
-                        
-            if (((current_round * 100) / this.rounds) % this.updateInterval == 0) {
-                this.progress = (current_round) * 100 / this.rounds;
-                WorkerEvent event;
-                
-                if (this.progress == 100) {
-                    this.buildWorkerResult();
-                    event = new WorkerEvent(WorkerEvent.EVENT_SIMWORKER_DONE, this.simResult);
-                    this.notifiable.onSimulationDone(event);
-                } else {
-                    event = new WorkerEvent(WorkerEvent.EVENT_SIMWORKER_PROGRESS, this.progress);
-                    this.notifiable.onSimulationProgress(event);
-                }
-            }
+
+            handleProgress(current_round);
+
         }
     }
 }
